+++ src/id_translation/fetching/__init__.py
@@ -16,6 +16,7 @@ Base fetchers:
       `placeholder mapping <../documentation/translation-primer.html#placeholder-mapping>`__.
 """
 from ._abstract_fetcher import AbstractFetcher
+from ._cache import CacheAccess, CacheMetadata
 from ._fetcher import Fetcher
 from ._memory_fetcher import MemoryFetcher
 from ._multi_fetcher import MultiFetcher
@@ -25,6 +26,8 @@ from ._sql_fetcher import SqlFetcher
 __all__ = [
     "Fetcher",
     "AbstractFetcher",
+    "CacheAccess",
+    "CacheMetadata",
     "MemoryFetcher",
     "MultiFetcher",
     "PandasFetcher",